* `Math:gen_rng` now has 2nd parameter `algorithm`
